+++ CHANGELOG.md
@@ -89,6 +89,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fixed pre-release archives issue. [#492]
+- Fixed mailto parsing issue. [core#10]
 
 ## [1.0.0-beta.4] - 2024-04-16
 
@@ -896,3 +897,4 @@ Few major concepts changed:
 
 [#492]: https://github.com/pimalaya/himalaya/issues/492
 [#496]: https://github.com/pimalaya/himalaya/issues/496
+[core#10]: https://github.com/pimalaya/core/issues/10